Scaleable editor programs. Enable dpi awareness

+++ src/cimgui_ext/cimgui_dock.h
@@ -13,7 +13,8 @@ extern "C" {
 #else
 #define IGEXPORT
 #endif
-IGEXPORT void *igFontFindGlyph(void *font, unsigned short c);
+#include <stdint.h>
+IGEXPORT void *igFontFindGlyph(void *font, uint32_t c);
 IGEXPORT bool igExtSplitterV(float thickness, float* size1, float *size2, float min_size1, float min_size2, float splitter_long_axis_size);
 #ifdef __cplusplus
 }
